hostmot2: Introduce queue_{read,write} methods
Some hostmot2 hardware, such as the ethernet-based 7i80, benefits
from batching reads and writes.  This will be done with the new queue
methods, queue_read and queue_write.

For reads, first a call to llio->queue_read with normal addr, buffer, and
size args are made for each read entry.  Then a final ->queue_read call
is made with addr=0, buffer=NULL, and args=-1.  The same happens for
writes.

Cards which do not benefit from batched writes need not implement the new
methods.  Instead, hm2_register will fill the slots with a function which
delegates to the traditional read method whenever the size argument is
not -1.