hostmot2: move watchdog I/O to the read function
On ethernet-connected mesa hardware, each packet sent or received
has an appreciable amount of overhead.  By performing the watchdog
read during the regular tram read process, the number of packets per
servo period is reduced by one.

The hal watchdog function still exists and must be added to a thread
in order to detect that the watchdog has bitten.